[Refactor] move weldx.tags.core.file.ExternalFile to weldx.util.external_file module
The implementation of ExternalFile is used in MediaFile, which is also living in the util module, and it seems better organized this way.

It was hard to convince Sphinx to reference ExternalFile, since "weldx.tags" is not part of the documentation.
